+++ CHANGELOG.md
@@ -4,6 +4,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- Add alpine js `clearable` component to input forms, use it by default in ig input component.
+
+### Fixed
+
+- Allow to define ig input class.
+
 ## [2.15.3] - 2025-10-14
 
 ### Fixed
@@ -830,6 +840,7 @@ _Stable release based on [0.1.0-rc.1]._
 
 ## [0.0.0] - 2024-09-12
 
+[Unreleased]: https://https://github.com/internetguru/laravel-common/compare/staging...dev
 [2.15.3]: https://https://github.com/internetguru/laravel-common/compare/v2.15.2...v2.15.3
 [2.15.2]: https://https://github.com/internetguru/laravel-common/compare/v2.15.1...v2.15.2
 [2.15.1]: https://https://github.com/internetguru/laravel-common/compare/v2.15.0...v2.15.1